split either and return as parser and validation errors
