+++ hmda/src/main/scala/hmda/api/http/model/filing/submissions/ValidationErrorSummary.scala
@@ -1,0 +1,6 @@
+package hmda.api.http.model.filing.submissions
+
+import hmda.model.validation.ValidationError
+
+case class ValidationErrorSummary(parserErrors: Seq[HmdaRowParsedErrorSummary],
+                                   validationErrors: Seq[Option[List[ValidationError]]])
